fixed code
Index 59 of year array is for leap day. I knew not to put data in it. However, I just incremented the counter and didn't put any value in after it was incremented. So it was adding in the wrong data for the next item in the array. It is now adding data to the right index in the array.