Completed graph one
This puts out the correct data for graph one. The line would only build to 59 and stop because of a null value. I fixed the array. The array adds the data correctly and averages the days correctly. I picked random days to check and average. I even checked the leap year day. Instead of months below its days.