Graph 1 & 2 Complete, tooltip works
Completed graph one and two. Now have a working tooltip on graph 2. The x-axis on graph two is also fixed. Had trouble with the legend for graph two. It's sensitive. Needed pad the legend with words to get births to appear in legend don't know why. Had to give the legend its own canvas, separate from the graph. Because the first bar in the bar graph would not show with original legend attach to main canvas.