Fixed transition of graphs
It's not a drop down menu. However, when a button is click it will remove any graph that is present and its legend. I also made the graphs align to the center of the page and its legend align to the center. No weird shift like in the before code. Where it was like stairs.